+++ src/widgets/post/index.ts
@@ -1,2 +1,3 @@
+export * from "./ui/PostAddDialog"
 export * from "./ui/PostDetailDialog"
 export * from "./ui/PostEditDialog"
